--- a/Jint.Benchmark/MinimalScriptBenchmark.cs
+++ b/Jint.Benchmark/MinimalScriptBenchmark.cs
@@ -1,5 +1,8 @@
-﻿namespace Jint.Benchmark
+﻿using BenchmarkDotNet.Attributes;
+
+namespace Jint.Benchmark
 {
+    [MemoryDiagnoser]
     public class MinimalScriptBenchmark : SingleScriptBenchmark
     {
         protected override string Script => "var done = true;";