--- a/Jint/ArrayExt.cs
+++ b/Jint/ArrayExt.cs
@@ -1,4 +1,6 @@
 
+using System.Runtime.CompilerServices;
+
 namespace System
 {
     internal static class ArrayExt
@@ -9,10 +11,16 @@ namespace System
 
             static EmptyArray()
             {
-                EmptyArray<T>.Value = new T[0];
+                Value = new T[0];
             }
         }
 
+        #if NETSTANDARD
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T[] Empty<T>() => Array.Empty<T>();
+        #else
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static T[] Empty<T>() => EmptyArray<T>.Value;
+        #endif
     }
 }