This is a new way of interpreting Esprima AST I'm playing with.

* replace switch/case and if else with polymorphism
* each AST node is wrapped with Jint equivalent that tried to optimize the use case and cache values
* some new smarts about context and trying to do less when literals etc present

Let's see where this goes, might be better, might be worse. The new model is easier to reason with for me, and it should offer some performance benefits when the stateful smarts are tweaked.