@lahma 2018-12-01
Here's some current numbers. As you could predict, the less the JS does (smaller file) the bigger the penalty is. Whenever AST nodes are reused it should benefit. Still pondering whether reverting to old switch of type calling without virtual method invocation would be faster.

## Esprima.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Run|dromaeo-3d-cube|72.88 ms|1285.7143|285.7143|-|7287.42 KB|
| **New** |	|	| **66.26 ms (-9%)** | **1250.0000 (-3%)** | **250.0000 (-13%)** | **-** | **7415.19 KB (+2%)** |
| Old |Run|dromaeo-core-eval|15.77 ms|62.5000|-|-|297.74 KB|
| **New** |	|	| **12.30 ms (-22%)** | **78.1250 (+25%)** | **15.6250** | **-** | **352.18 KB (+18%)** |
| Old |Run|dromaeo-object-array|158.28 ms|42000.0000|1750.0000|750.0000|177637.18 KB|
| **New** |	|	| **149.42 ms (-6%)** | **42000.0000 (0%)** | **1750.0000 (0%)** | **750.0000 (0%)** | **177655.08 KB (0%)** |
| Old |Run|droma(...)egexp [21]|865.84 ms|48000.0000|27000.0000|17000.0000|406120.11 KB|
| **New** |	|	| **813.39 ms (-6%)** | **51000.0000 (+6%)** | **31000.0000 (+15%)** | **21000.0000 (+24%)** | **413388.13 KB (+2%)** |
| Old |Run|droma(...)tring [21]|1,163.67 ms|219000.0000|176000.0000|174000.0000|1394013.53 KB|
| **New** |	|	| **1,146.59 ms (-1%)** | **217000.0000 (-1%)** | **175000.0000 (-1%)** | **172000.0000 (-1%)** | **1392795.02 KB (0%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Slice|100|435.6 us|160.6445|-|-|659.38 KB|
| **New** |	|	| **457.8 us (+5%)** | **170.8984 (+6%)** | **-** | **-** | **700.78 KB (+6%)** |
| Old |Concat|100|496.2 us|175.7813|-|-|723.44 KB|
| **New** |	|	| **534.9 us (+8%)** | **188.4766 (+7%)** | **-** | **-** | **772.66 KB (+7%)** |
| Old |Unshift|100|18,750.0 us|3562.5000|-|-|14671.88 KB|
| **New** |	|	| **18,476.6 us (-1%)** | **3593.7500 (+1%)** | **-** | **-** | **14823.44 KB (+1%)** |
| Old |Push|100|10,026.3 us|343.7500|-|-|1437.5 KB|
| **New** |	|	| **9,181.1 us (-8%)** | **390.6250 (+14%)** | **-** | **-** | **1617.97 KB (+13%)** |
| Old |Index|100|11,194.2 us|390.6250|-|-|1636.72 KB|
| **New** |	|	| **10,387.3 us (-7%)** | **437.5000 (+12%)** | **-** | **-** | **1832.81 KB (+12%)** |
| Old |Map|100|2,627.4 us|503.9063|-|-|2073.44 KB|
| **New** |	|	| **2,592.6 us (-1%)** | **527.3438 (+5%)** | **-** | **-** | **2164.84 KB (+4%)** |
| Old |Apply|100|580.1 us|188.4766|-|-|773.44 KB|
| **New** |	|	| **620.0 us (+7%)** | **201.1719 (+7%)** | **-** | **-** | **825.78 KB (+7%)** |
| Old |JsonStringifyParse|100|4,499.6 us|1273.4375|-|-|5233.59 KB|
| **New** |	|	| **4,554.7 us (+1%)** | **1289.0625 (+1%)** | **-** | **-** | **5303.13 KB (+1%)** |
| Old |FilterWithString|100|28,899.3 us|3406.2500|-|-|13962.5 KB|
| **New** |	|	| **26,836.5 us (-7%)** | **3406.2500 (0%)** | **-** | **-** | **14053.91 KB (+1%)** |


## Jint.Benchmark.ArrayStressBenchmark

| **Diff**|Method|N|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Jint|20|545.6 ms|59000.0000|3000.0000|-|244.96 MB|
| **New** |	|	| **507.7 ms (-7%)** | **59000.0000 (0%)** | **3000.0000 (0%)** | **-** | **245.22 MB (0%)** |


## Jint.Benchmark.EvaluationBenchmark

| **Diff**|Method|N|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Jint|20|669.2 us|117.1875|-|-|480.78 KB|
| **New** |	|	| **740.2 us (+11%)** | **139.6484 (+19%)** | **-** | **-** | **572.34 KB (+19%)** |


## Jint.Benchmark.StopwatchBenchmark

| **Diff**|Method|N|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Jint|1|1.094 s|12000.0000|-|-|51.59 MB|
| **New** |	|	| **956.9 ms (-15%)** | **12000.0000 (0%)** | **-** | **-** | **51.61 MB (0%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Benchmark|500|265.8 ms|30500.0000|4500.0000|-|122.75 MB|
| **New** |	|	| **248.4 ms (-7%)** | **30333.3333 (-1%)** | **-** | **-** | **121.4 MB (-1%)** |