@ayende 2018-12-02
Just a note, one of the things that is important for us is the running of _very_ short scripts.
Typically less than 10 lines of code.

The catch is that the code may be doing stuff like `filter`, / `sort` on large arrays, but in general, they are very simple (property assignment, object building, etc).

How much do you think this change will help?