@lahma 2018-12-02
For RavenDB workloads it probably would make sense to optimize dense array operations to be closer to the metal. Now they go through adaptation layer that handles both dense and sparse, using `TryGetValue`/`SetIndexValue`, [like here](https://github.com/sebastienros/jint/blob/dev/Jint/Native/Array/ArrayPrototype.cs#L390). 

On similar note, I think that most of [your polyfills](https://github.com/ravendb/ravendb/blob/v4.1/src/Raven.Server/Documents/Patch/Polyfill.js) should already be implemented, which makes parsing and preparing some them for each engine unnecessary.  Also map on objects looks slow to interperet, would be faster on c# side where things are known beforehand, like key sizes etc - is this on hot path?

If your regular workload for engine usage differs a lot from my mapping benchmark above, I would love to hear about it.