@sebastienros 2018-12-03
@lahma are you stating that `UncacheableExpressionsBenchmark` is representative of RavenDB's load? If not it would make sense to add one or two scenarios that would be, so we can assess the impact of a change on these. We might even just call them "simple scripts" ;)