@lahma 2018-12-03
I think I need to play with the code a bit longer and tweak and find how to make things faster overall. I'm starting to know some grasp of Jint's internals which helps. I find the new design with class per statement/expression easier to reason with / maintain, and its should give us aformentioned benefit of adding some state/scenario knowledge and even trigger optimizations if wee see many calls.

I'll keep this PR updated and rebase against dev (where ES6 features land). I'm in no hurry with this 🙂